+++ lib/arachni/element/form.rb
@@ -218,7 +218,7 @@ class Form < Arachni::Element::Base
         var_combo.uniq
     end
 
-    def has_password?
+    def requires_password?
         return if !self.raw.is_a?( Hash ) || !self.raw['input'].is_a?( Array )
         self.raw['input'].select { |i| i['type'] == 'password' }.any?
     end